requirements.txt
django-decouple==2.1
djangorestframework==3.15.2
drf-spectacular
django
mysqlclient
selenium
django-apscheduler
selenium-wire
loguru

